put union method in the right place
