+++ nsdb-core/src/main/scala/io/radicalbit/nsdb/model/Schema.scala
@@ -20,7 +20,7 @@ import io.radicalbit.nsdb.common.JSerializable
 import io.radicalbit.nsdb.common.protocol.{Bit, FieldClassType}
 import io.radicalbit.nsdb.index.{IndexType, TypeSupport}
 
-import scala.util.Try
+import scala.util.{Failure, Success, Try}
 
 /**
   * It models a bit field
@@ -75,5 +75,29 @@ object Schema extends TypeSupport {
   def apply(metric: String, bit: Bit): Try[Schema] = {
     validateSchemaTypeSupport(bit).map((fields: Seq[TypedField]) =>
       Schema(metric, fields.map(field => SchemaField(field.name, field.fieldClassType, field.indexType)).toSet))
+  }
+
+  /**
+    * Assemblies, if possible, the union schema from 2 given schemas.
+    * Given 2 schemas, they are compatible if fields present in both of them are of the same types.
+    * The union schema is a schema with the union of the dimension sets.
+    * @param firstSchema the first schema.
+    * @param secondSchema the second schema.
+    * @return the union schema.
+    */
+  def union(firstSchema: Schema, secondSchema: Schema): Try[Schema] = {
+    val oldFields = firstSchema.fields.map(e => e.name -> e).toMap
+
+    val notCompatibleFields = secondSchema.fields.collect {
+      case field if oldFields.get(field.name).isDefined && oldFields(field.name).indexType != field.indexType =>
+        s"mismatch type for field ${field.name} : new type ${field.indexType} is incompatible with old type"
+    }
+
+    if (notCompatibleFields.nonEmpty)
+      Failure(new RuntimeException(notCompatibleFields.mkString(",")))
+    else {
+      val schema = Schema(secondSchema.metric, firstSchema.fields ++ secondSchema.fields)
+      Success(schema)
+    }
   }
 }
